:truck: Move Generic Card into plugin module
