+++ src/frontend/modules/plugins/components/PluginCard/GenericCard.tsx
