+++ src/frontend/modules/plugins/components/PluginCard/PluginCard.tsx
@@ -1,11 +1,11 @@
 import type { PluginApi } from "@/api/fastapi-schemas";
 import { useDeletePlugin } from "@/api/fastapi/plugins/plugins";
-import { GenericCard } from "@/components/Cards/GenericCard";
 import UploadModal from "@/components/UploadModal/UploadModal";
 import getModuleRoute from "@/lib/modules/getModuleRoute";
 import { Card, Menu, Stack, Text, ThemeIcon } from "@mantine/core";
 import { Trash2Icon, UploadIcon } from "lucide-react";
 import { useState } from "react";
+import { GenericCard } from "./GenericCard";
 
 export const PluginCard: React.FC<{ plugin: PluginApi }> = ({ plugin }) => {
   const { description, label, version } = plugin.meta;
